Add the FEC regeneration for D-Star, DMR, and P25.

+++ Codec23200PCM.h
@@ -26,7 +26,7 @@ class CCodec23200PCM : public IProcessor {
     CCodec23200PCM();
     virtual ~CCodec23200PCM();
 
-    virtual uint8_t  input(const uint8_t* buffer, uint8_t length);
+    virtual uint8_t  input(const uint8_t* buffer, uint16_t length);
 
     virtual uint16_t output(uint8_t* buffer);
