+++ YSFDNDMRNXDN.h
@@ -26,7 +26,7 @@ class CYSFDNDMRNXDN : public IProcessor {
     CYSFDNDMRNXDN();
     virtual ~CYSFDNDMRNXDN();
 
-    virtual uint8_t  input(const uint8_t* buffer, uint8_t length);
+    virtual uint8_t  input(const uint8_t* buffer, uint16_t length);
 
     virtual uint16_t output(uint8_t* buffer);
